Fixed 'Union' method in 'Network.js'
Problem: My Union function was causing nodes
of the same components to establish new direct
connection with each other.

Cause: Using my 'test.js', I noticed that
I was making the "node" part of the smaller
component point to the one in the bigger
component.

Fix: Make the "root node" of the smaller
component point to the "root node" of the
bigger component.